omxvideodec: change supported color formats list creation
To create supported color formats list for caps negotiation,
this change tries to set each color format that is prepared in
a list beforehand and determines which color formats are supported
by the component. The more components can be dealt with by this
support.

NV12 color format is put at the top of the list declared as a constant
so that the caps negotiation gives top priority to this color format.