omxvideodec: don't use gst_buffer_map when copying to the output buffers
When the preallocated buffers which are separated into multiple blocks
are received from a sink plugin, gst_buffer_map() could return
the different pointer from the buffers received from the sink plugin.
Probably it is newly allocated because the buffers are mistakenly
regarded as being discontiguous and tried to be collected into a
region with copying.
It is safety that the stride and the slice height are always taken
account into when copying to the output buffers.