omxh264dec: Support the frame unit input into REL OMXIL
This change would alleviate the overhead due to input stream
copying. The OMXIL provided by REL recommends that
OMXR_MC_VIDEO_StoreUnitEofSeparated, which is set to the OMXIL
by default, is used rather than OMXR_MC_VIDEO_StoreUnitTimestampSeparated.

The OMXIL is requried to receive the bytestream data format in a frame,
so the patch transforms the AVC sample, which is the format that
is applied to the H.264 stream in the mp4 container, into the bytestream.