omxvideodec: Don't release the buffer that haven't been acquired yet
This patch fixes a problem with handling the gst-omx buffer pool.

The buffer pool attempts to insert the output buffers from AllocateBuffer()
into the omxil component when the buffer pool is activated in dec_loop().
Also set_format(), which is invoked before dec_loop(), sends
the output buffers into the component with gst_omx_port_populate(),
which causes a malfunction in the buffer management for the duplicated
buffers insertion.

However the output buffer insertion into the omx component in set_format()
is necessary because the caps which the buffer pool needs to be configured
is determined after the first frame has been decoded, which can be noticed
from the output frame acquisition processing in dec_loop().

This patch uses the already_acquired flag to avoid the duplicated
buffers insertion into the component. The flag is set once
the buffer is obtained by gst_buffer_pool_acquire_buffer() in dec_loop()
and allows to release the buffer back to the output port in buffer_release()
only after the acquisition.