omxvideodec: Fix the buffer index determination to acquire a buffer from the buffer pool
The GstOMXBuffer instances are associated with the buffers queued
in the output buffer pool. This index determination processing is
implemented on the premise that the GstOMXBuffer instances queued
in port->buffers is the same order with the ones queued in
the output buffer pool. If the order of the either queue is changed,
the processing can't correctly perform.

Actually the index specified by this determination is used in
gst_omx_buffer_pool_acquire_buffer() in order to obtain a buffer
from the output buffer pool, so the determinaion would rather see
the queue of the buffer pool than port->buffers.