omxvideodec: Associate dmabuf file descriptors with output buffers when the mmngrbuf library is detected
This patch exports dmabuf file descriptors with the MMP mmngrbuf API
and attach them to output buffers.

To export dmabuf file descriptors, the physical address which is
obtained from the decoding result of each frame has to be passed
into the dmabuf exporter API.

A new GstBuffer which has no GstMemory instances is created
when alloc_buffer() is invoked, then a dmabuf file descriptor is
attached to the GstBuffer when acquire_buffer() is invoked
because the phycical address can be obtained after invoking
the FillBufferDone callback for each frame.