omxvideodec: Workaround the timestamp issue with h264parse
In GStreamer, there are two timestamps for each frame data.
They are pts and dts. gst-omx uses pts.

In GStreamer 1.2.3, for an unknown reason, h264parse does not send pts for
some videos, such as RenesasElectronicsVGA.264. In old Gstreamer (1.0.7) this
does not happen

Since h264parse does not send pts, gst-omx cannot get the timestamp for buffer
and as the result, the display video is very lag.

refs #54088
Signed-off-by: Thao Luong <thao.luong.uw@rvc.renesas.com>