omxvideodec: Activate the buffer pool of gst-omx by itself rather than by the gstvideodecoder
This change removes a reference to the buffer pool of gst-omx
from the gstvideodecoder as a preliminary modification in order to
fix the memory leak due to two way reference between the buffer pool
and the gstvideodecoder.
The buffer pool of gst-omx had been controlled regarding only
the state transition by the gstvideodecoder, so this change does not
set the buffer pool to the gstvideodecoder and invokes activation of
the buffer pool in this plugin.
The deactivation of the buffer pool has already been called in
gst_omx_video_dec_deallocate_output_buffers().

The video buffer pool, which is the default buffer pool in the GStreamer
video library, is set to the gstvideodecoder instead, because
it is mandatoty to set a buffer pool into the gstvideodecoder class.