Revert "omxvideodec: Fix the duplicated dereferencing of our buffer pool"
This reverts commit b40a656a14b5ef9cf4da048e3a857ddcb9a435ab.

Before this change, it was expected to do dereferencing of our buffer pool
by the gstvideodecoder class, but that led to two way referencing
between the buffer pool and the gstvideodecoder class. Consequently
a memory leak of our buffer pool had occured.

The reversion is necessary to do dereferencing of our buffer pool
by ourself, avoiding the two way referencing.