omxvideodec: Drop frames which are not keyframes in the backward playback
The backward playback algorithm in GStreamer sorts decoded frames
in reverse order for each GOP, so the number of output buffers as
the same as GOP are needed.

This change achieves the backward playback without resort to the
algorithm by outputting only keyframes and dropping others,
not doing the reordering in gst-omx as the demuxer moves the
read position backward in steps of a keyframe.